Separate contract class files, rename builtinSources
