+++ builtin-contracts/src/main/java/com/aws/jverify/builtin/ComparableContract.java
@@ -1,0 +1,7 @@
+package com.aws.jverify.builtin;
+
+import com.aws.jverify.Contract;
+
+@Contract(Comparable.class)
+class ComparableContract<T> {
+}
