+++ builtin-contracts/src/main/java/com/aws/jverify/builtin/IntegerContractHelper.java
@@ -1,0 +1,16 @@
+package com.aws.jverify.builtin;
+
+import com.aws.jverify.Pure;
+
+class IntegerContractHelper {
+    @Pure
+    public static boolean sumWillOverflow(int a, int b) {
+        if (a > 0 && b > 0) {
+            return a > Integer.MAX_VALUE - b;
+        }
+        if (a < 0 && b < 0) {
+            return a < Integer.MIN_VALUE - b;
+        }
+        return false; // Different signs won't overflow
+    }
+}
