+++ builtin-contracts/src/main/java/com/aws/jverify/builtin/LongContract.java
@@ -1,0 +1,23 @@
+package com.aws.jverify.builtin;
+
+import com.aws.jverify.Contract;
+import com.aws.jverify.ContractException;
+import com.aws.jverify.Pure;
+import static com.aws.jverify.JVerify.*;
+
+@Contract(Long.class)
+class LongContract {
+    public static final long MIN_VALUE = 0x8000000000000000L;
+    public static final long MAX_VALUE = 0x7fffffffffffffffL;
+
+    @Pure
+    public static Long valueOf(long l) {
+        postcondition((Long b) -> b.longValue() == l);
+        throw new ContractException();
+    }
+
+    @Pure
+    public long longValue() {
+        throw new ContractException();
+    }
+}
