+++ builtin-contracts/src/main/java/com/aws/jverify/builtin/MapEntryContract.java
@@ -1,0 +1,21 @@
+package com.aws.jverify.builtin;
+
+import com.aws.jverify.Contract;
+import com.aws.jverify.ContractException;
+import com.aws.jverify.Pure;
+import java.util.Map;
+
+@Contract(pure = true)
+abstract class MapEntryContract<K, V> implements Map.Entry<K, V> {
+    @Override
+    @Pure
+    public K getKey() {
+        throw new ContractException();
+    }
+
+    @Override
+    @Pure
+    public V getValue() {
+        throw new ContractException();
+    }
+}
