+++ builtin-contracts/src/main/java/com/aws/jverify/builtin/NumberContract.java
@@ -1,0 +1,8 @@
+package com.aws.jverify.builtin;
+
+import com.aws.jverify.Contract;
+
+@Contract(value = Number.class, pure = true)
+class NumberContract {
+
+}
