+++ builtin-contracts/src/main/java/com/aws/jverify/builtin/PredicateContract.java
@@ -1,0 +1,8 @@
+package com.aws.jverify.builtin;
+
+import com.aws.jverify.Contract;
+import java.util.function.Predicate;
+
+@Contract
+abstract class PredicateContract<T> implements Predicate<T> {
+}
