+++ builtin-contracts/src/main/java/com/aws/jverify/builtin/SequencedCollectionContract.java
@@ -1,0 +1,8 @@
+package com.aws.jverify.builtin;
+
+import com.aws.jverify.Contract;
+import java.util.Collection;
+import java.util.SequencedCollection;
+
+@Contract(SequencedCollection.class)
+interface SequencedCollectionContract<E> extends Collection<E> { }
