+++ builtin-contracts/src/main/java/com/aws/jverify/builtin/SystemContract.java
@@ -1,0 +1,10 @@
+package com.aws.jverify.builtin;
+
+import com.aws.jverify.Contract;
+import java.io.PrintStream;
+
+@Contract(System.class)
+class SystemContract {
+    public static final PrintStream out = SystemContractHelper.createPrintStream();
+    
+}
