+++ builtin-contracts/src/main/java/com/aws/jverify/builtin/SystemContractHelper.java
@@ -1,0 +1,14 @@
+package com.aws.jverify.builtin;
+
+import com.aws.jverify.ContractException;
+import com.aws.jverify.Pure;
+import com.aws.jverify.Verify;
+import java.io.PrintStream;
+
+class SystemContractHelper {
+    @Verify(false)
+    @Pure
+    public static PrintStream createPrintStream() {
+        throw new ContractException();
+    }
+}
