+++ verifier/src/test/java/com/aws/jverify/verifier/tests/jcl/CollectionsTest.java
@@ -1,4 +1,4 @@
-// ^ /aws/jverify/builtin/builtin-contracts.java(171:36-171:50) Related location: this proposition could not be proved
+// ^ /aws/jverify/builtin/ListContract.java(48:36-48:50) Related location: this proposition could not be proved
 package com.aws.jverify.verifier.tests.jcl;
 
 import com.aws.jverify.testengine.JVerifyTest;
