+++ verifier/src/test/java/com/aws/jverify/verifier/tests/jcl/Optionals.java
@@ -1,4 +1,4 @@
-// ^ /aws/jverify/builtin/builtin-contracts.java(736:22-736:27) Related location: this proposition could not be proved
+// ^ /aws/jverify/builtin/OptionalContract.java(57:22-57:27) Related location: this proposition could not be proved
 package com.aws.jverify.verifier.tests.jcl;
 
 import com.aws.jverify.testengine.JVerifyTest;
